… CHANGELOG.rst
… tests/mapping/test_BaseMappingTransformMixin.py
… tests/mapping/test_MappingTransformer.py
… tests/nominal/test_MeanResponseTransformer.py
… tests/nominal/test_NominalToIntegerTransformer.py
… tests/nominal/test_OrdinalEncoderTransformer.py
… tubular/nominal.py