@liamholmes31 2025-08-05
Some info from the profiling (ran on 3 columns to get some different types/scenarios in):
- original (non narwhals tubular) - ~0.0088s
- pre optimised narwhals tubular + pandas - ~0.037s
- optimised narwhals tubular + pandas - ~0.022s
- optimised narwhals tubular + polars - ~0.0056s

so we are still taking a hit short term when working with pandas, but hopefully this work helps to keep it manageable until we move to polars